+++ src/LageBuch.AppLogic/ViewModels/RolesViewModel.cs
@@ -1,7 +1,7 @@
 using System.Collections.ObjectModel;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
-using LageBuch.Documents;
+using LageBuch.Domain;
 using LageBuch.Domain.Time;
 using LageBuch.Persistence.MasterData;